set sessionExpiryInterval to long value in WS connection configuration to be able to support its max value

+++ common/data/src/main/java/org/thingsboard/mqtt/broker/common/data/ws/WebSocketConnectionConfiguration.java
@@ -44,7 +44,7 @@ public class WebSocketConnectionConfiguration {
     private int reconnectPeriod;
     private TimeUnit reconnectPeriodUnit;
     private int mqttVersion;
-    private int sessionExpiryInterval;
+    private long sessionExpiryInterval;
     private TimeUnit sessionExpiryIntervalUnit;
     private int maxPacketSize;
     private SizeUnit maxPacketSizeUnit;
